apple-silicon-plugin/detector/requirements.txt
onnxruntime-silicon>=1.16.0
pyzmq>=25.1.0
numpy>=1.24.0
opencv-python>=4.8.0
pyyaml>=6.0
